+++ sandbox/auth-ms-basic-example/src/repositories/index.ts
@@ -3,3 +3,6 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 export * from './to-do.repository';
+export * from './role.repository';
+export * from './tenant.repository';
+export * from './user.repository';