+++ services/audit-service/src/enums/permission-key.enum.ts
@@ -7,4 +7,8 @@ export const enum PermissionKey {
   CreateAudit = 'CreateAudit',
   UpdateAudit = 'UpdateAudit',
   DeleteAudit = 'DeleteAudit',
+  ViewAuditNum = '1',
+  CreateAuditNum = '2',
+  UpdateAuditNum = '3',
+  DeleteAuditNum = '4',
 }