+++ services/chat-service/src/controllers/message-recipient-message.controller.ts
@@ -24,7 +24,9 @@ export class MessageRecipientMessageController {
   ) {}
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKey.ViewMessage]})
+  @authorize({
+    permissions: [PermissionKey.ViewMessage, PermissionKey.ViewMessageNum],
+  })
   @get(basePath, {
     security: OPERATION_SECURITY_SPEC,
     responses: {