+++ services/chat-service/src/enums/permission-key.enum.ts
@@ -12,4 +12,18 @@ export enum PermissionKey {
   ViewMessageRecipient = 'ViewMessageRecipient',
   UpdateMessageRecipient = 'UpdateMessageRecipient',
   DeleteMessageRecipient = 'DeleteMessageRecipient',
+
+  ViewMessageNum = '1',
+  CreateMessageNum = '2',
+  UpdateMessageNum = '3',
+  DeleteMessageNum = '4',
+
+  CreateMessageRecipientNum = '5',
+  ViewMessageRecipientNum = '6',
+  UpdateMessageRecipientNum = '7',
+  DeleteMessageRecipientNum = '8',
+  CreateAttachmentFile = 'CreateAttachmentFile',
+  ViewAttachmentFile = 'ViewAttachmentFile',
+  UpdateAttachmentFile = 'UpdateAttachmentFile',
+  DeleteAttachmentFile = 'DeleteAttachmentFile',
 }