+++ services/scheduler-service/src/controllers/event.controller.ts
@@ -62,7 +62,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.CreateEvent]})
+  @authorize({
+    permissions: [PermissionKey.CreateEvent, PermissionKey.CreateEventNum],
+  })
   @post(basePath, {
     description: `While an organizer creates an event, we get participants details along with it. 
       This api will check for slot availability of all the participants in that particular 
@@ -133,7 +135,12 @@ export class EventController {
     passReqToCallback: true,
   })
   @authorize({
-    permissions: [PermissionKey.ViewEvent, PermissionKey.ViewAttendee],
+    permissions: [
+      PermissionKey.ViewEvent,
+      PermissionKey.ViewAttendee,
+      PermissionKey.ViewEventNum,
+      PermissionKey.ViewAttendeeNum,
+    ],
   })
   @get('/events/freeBusy', {
     security: OPERATION_SECURITY_SPEC,
@@ -196,7 +203,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.ViewEvent]})
+  @authorize({
+    permissions: [PermissionKey.ViewEvent, PermissionKey.ViewEventNum],
+  })
   @get(`${basePath}/count`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
@@ -213,7 +222,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.ViewEvent]})
+  @authorize({
+    permissions: [PermissionKey.ViewEvent, PermissionKey.ViewEventNum],
+  })
   @get(basePath, {
     description:
       'This api will return the events data, based on the filter provided. Sending the data of participants will be optional and will depend on the query.',
@@ -254,7 +265,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.UpdateEvent]})
+  @authorize({
+    permissions: [PermissionKey.UpdateEvent, PermissionKey.UpdateEventNum],
+  })
   @patch(basePath, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
@@ -281,7 +294,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.ViewEvent]})
+  @authorize({
+    permissions: [PermissionKey.ViewEvent, PermissionKey.ViewEventNum],
+  })
   @get(`${basePath}/{id}`, {
     description:
       'This api will return events data based on the id. Sending the data of participants will be optional and will depend on the query.',
@@ -308,7 +323,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.UpdateEvent]})
+  @authorize({
+    permissions: [PermissionKey.UpdateEvent, PermissionKey.UpdateEventNum],
+  })
   @patch(`${basePath}/{id}`, {
     description:
       'This api will be responsible for making any updates on an event. This action is only allowed to the organizer or the admin(based on permission).',
@@ -336,7 +353,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.UpdateEvent]})
+  @authorize({
+    permissions: [PermissionKey.UpdateEvent, PermissionKey.UpdateEventNum],
+  })
   @put(`${basePath}/{id}`, {
     description:
       'This api will be responsible for making any updates on an event. This action is only allowed to the organizer or the admin(based on permission).',
@@ -357,7 +376,9 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.DeleteEvent]})
+  @authorize({
+    permissions: [PermissionKey.DeleteEvent, PermissionKey.DeleteEventNum],
+  })
   @del(`${basePath}/{id}`, {
     description:
       'Api to delete the event based on id. The action is only allowed to the organiser or the admin(based on permission).',
@@ -379,7 +400,12 @@ export class EventController {
   @authenticate(STRATEGY.BEARER, {
     passReqToCallback: true,
   })
-  @authorize({permissions: [PermissionKey.HardDeleteEvent]})
+  @authorize({
+    permissions: [
+      PermissionKey.HardDeleteEvent,
+      PermissionKey.HardDeleteEventNum,
+    ],
+  })
   @del(`${basePath}/{id}/hard`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {