+++ services/scheduler-service/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>5.2.7 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+## <small>5.2.6 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+## <small>5.2.5 (2022-09-21)</small>
+
+* chore(deps): update loopback4-soft-delete version in services (#1030) ([d386bc1](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/d386bc1)), closes [#1030](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1030) [#1026](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1026)
+
+
+
+
+
+## <small>5.2.4 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+## <small>5.2.3 (2022-09-20)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+## <small>5.2.2 (2022-09-12)</small>
+
+* chore(deps): loopback version update for services (#1017) ([9a9d653](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/9a9d653)), closes [#1017](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1017) [#1015](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1015)
+
+
+
+
+
+## <small>5.2.1 (2022-09-12)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+## 5.2.0 (2022-09-06)
+
+* feat(all-services): created number as string for permissions (#1006) ([396121b](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/396121b)), closes [#1006](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1006) [#994](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/994)
+
+
+
+
+
 ## <small>5.1.1 (2022-09-02)</small>
 
 * chore(deps): fixing typescript version for cli (#1005) ([9e561f2](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/9e561f2)), closes [#1005](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1005) [#00](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/00)